Mix the actor actions from oot and mm
